+++ Thunderstore/CHANGELOG.md
@@ -3,6 +3,9 @@
 
 * Witch Hunter's Tools
 	* Removed unneeded SphereCollider so it should no longer have collision.
+* Adrenaline Core
+	* Added Transcendence check so if player has it, item checks for shield damage instead of health damage. This can be disabled in the config.
+	* Added additional checks to health, so when player loses health from scrapping items and other things that modify max health, delevel won't trigger.
 * Fixed sound attenuation for all items and equipment, it should now be 100m for everything instead of being 30m for some and infinite for others.
 * Removed some BetterUI tags from items so they no longer "stack" stats with other items.
 * Added missing R2API Language and Prefab dependency.